requirements.txt
fastapi
uvicorn
jinja2
python-multipart
supabase
google-generativeai
pydantic
python-dotenv
